requirements.txt
streamlit>=1.30
opencv-python-headless>=4.8
numpy>=1.24
scipy>=1.10
pillow>=10
